Move HF source tests
